almost norminetted, somethings broken rn, but i have to leave lol
