+++ src/parser/files.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   files.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: plouda <plouda@student.42.fr>              +#+  +:+       +#+        */
+/*   By: ehasalu <ehasalu@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/01 11:44:04 by plouda            #+#    #+#             */
-/*   Updated: 2023/08/29 11:38:12 by plouda           ###   ########.fr       */
+/*   Updated: 2023/09/01 17:04:05 by ehasalu          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,23 +14,10 @@
 
 // Should probably not pass a whole stucture if it can be avoided
 
-void	here_doc(char **argv, int i, t_command *command)
+void	here_doc2(char **argv, int i, t_command *command, char **delimiter_cpy)
 {
-	int		j;
-	char	**delimiter_cpy;
+	int	j;
 
-	command->here_doc = HERE_DOC_IN;
-	if (!command->valid)
-		command->here_doc = HERE_DOC_VOID;
-	delimiter_cpy = malloc(sizeof(char *) * (command->here_doc_counter + 1));
-	j = 0;
-	while (command->delimiter != NULL && j < command->here_doc_counter)
-	{
-		delimiter_cpy[j] = ft_strdup(command->delimiter[j]);
-		j++;
-	}
-	delimiter_cpy[j] = NULL;
-	command->here_doc_counter++;
 	j = 0;
 	if (command->delimiter != NULL)
 	{
@@ -54,6 +41,26 @@ void	here_doc(char **argv, int i, t_command *command)
 	free(delimiter_cpy);
 	command->delimiter[j++] = ft_strdup(argv[i + 1]);
 	command->delimiter[j] = NULL;
+}
+
+void	here_doc(char **argv, int i, t_command *command)
+{
+	int		j;
+	char	**delimiter_cpy;
+
+	command->here_doc = HERE_DOC_IN;
+	if (!command->valid)
+		command->here_doc = HERE_DOC_VOID;
+	delimiter_cpy = malloc(sizeof(char *) * (command->here_doc_counter + 1));
+	j = 0;
+	while (command->delimiter != NULL && j < command->here_doc_counter)
+	{
+		delimiter_cpy[j] = ft_strdup(command->delimiter[j]);
+		j++;
+	}
+	delimiter_cpy[j] = NULL;
+	command->here_doc_counter++;
+	here_doc2(argv, i, command, delimiter_cpy);
 }
 
 void	infile(char **argv, int i, t_command *command)
